Update contollers and view/shared files
