+++ app/assets/stylesheets/application.css
@@ -13,3 +13,132 @@
  *= require_tree .
  *= require_self
  */
+
+.users {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  align-items: center;
+  justify-content: center;
+  flex-basis: 500px;
+}
+
+.users a {
+  width: 500px;
+  text-decoration: none;
+  color: blue;
+}
+
+.posts a {
+  text-decoration: none;
+  color: rgb(51, 51, 53);
+}
+
+.users a:hover {
+  text-decoration: underline;
+  color: red;
+}
+
+.users a:visited {
+  color: rgb(40, 40, 41);
+}
+
+.user-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  border: 2px solid #232222;
+  border-radius: 5px;
+  padding: 1rem;
+  height: 200px;
+  box-shadow: -3px 15px 5px -10px rgba(0, 0, 0, 0.67);
+  -webkit-box-shadow: -3px 15px 5px -10px rgba(0, 0, 0, 0.67);
+  -moz-box-shadow: -3px 15px 5px -10px rgba(0, 0, 0, 0.67);
+}
+
+.img-container {
+  width: 100px;
+  height: 100px;
+}
+
+.img-container img {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+}
+
+.show-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+}
+
+.posts {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 100vw;
+  align-items: center;
+}
+
+.post {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 500px;
+  padding: 1rem;
+  border: 2px solid #232222;
+}
+
+.interactions {
+  display: flex;
+  gap: 1rem;
+  width: 100%;
+  justify-content: right;
+}
+
+.post-title {
+  display: flex;
+  width: 100%;
+  justify-content: left;
+}
+
+.comments {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  border: 1px solid #232222;
+  padding: 1rem;
+  width: 500px;
+}
+
+.posts-show {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  align-items: center;
+  width: 100vw;
+}
+
+.post-show {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 500px;
+  padding: 1rem;
+  border: 2px solid #232222;
+}
+
+.post-title-show {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.post-interact {
+  display: flex;
+  gap: 1rem;
+}
