+++ app/controllers/concerns/users_controller.rb
@@ -1,15 +1,15 @@
 class UsersController < ApplicationController
   def index
-  @users = User.all
+    @users = User.all
   end
 
   def show
-  @user = User.find_by_id(params[:id])
-  if @user
-  @recent_posts = @user.most_recent_posts
-  else
-    flash[:alert] = 'User not found'
-    redirect_to root_path
+    @user = User.find_by_id(params[:id])
+    if @user
+      @recent_posts = @user.most_recent_posts
+    else
+      flash[:alert] = 'User not found'
+      redirect_to root_path
+    end
   end
-end
 end
