+++ app/helpers/likes_helper.rb
@@ -1,0 +1,2 @@
+module LikesHelper
+end
