+++ app/views/likes/like.html.erb
@@ -1,6 +1,5 @@
 <% stylesheet_link_tag('index')%>
 <div class="likes-container">
-<%# <h3>Add a new Comment</h3> %>
 <button class="Add-likes">Like comment/post</button>
 <div class="comment-form">
 </div>
