+++ app/views/posts/index.html.erb
@@ -1,4 +1,3 @@
-<%= stylesheet_link_tag('index') %>
 <div class="posts-container">
 <div class="users">
 <%= render 'shared/user', user:@user %>
