+++ app/views/posts/show.html.erb
@@ -1,4 +1,3 @@
-<%= stylesheet_link_tag('index') %>
 <div class="posts-show">
 <div class="post-show">
