+++ app/views/users/show.html.erb
@@ -1,18 +1,13 @@
-<div class="post-container">
-<div class="post-heading">
+<div class="show-page">
+<div class="users">
 <%= render 'shared/user', user:@user %>
 </div>
 <div class="posts">
-<h2>Username</h2>
+<h2>Recent Posts</h2>
 <% @recent_posts.each do |post|
  %>
 <%= render 'shared/post', post:post, user:@user %>
 <% end %>
-<%# <div class="btn">
-    <button>
-      <%= link_to 'Create new post', new_user_post_path(current_user) %>
-    <%# </button> %>
-  <%# </div>   %>
 </div>
 <%=link_to 'See all posts', user_posts_path(@user) %>
 </div>
